pyproject.toml
[build-system]
requires = ["uv_build>=0.8.5,<0.9.0"]
build-backend = "uv_build"

[project]
name = "zotero-arxiv-daily"
version = "0.3.5"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gitignore-parser>=0.1.11",
    "tiktoken>=0.8.0",
    "feedparser>=6.0.11",
    "hydra-core>=1.3.2",
    "aiosmtpd>=1.4.6",
    "loguru>=0.7.3",
    "pyzotero>=1.6.11",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=5.0.0",
    "openai>=1.99.1",
    "llama-cpp-python>=0.3.2",
    "arxiv>=2.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
